requirements.txt
streamlit>=1.28.0,<2.0.0
anthropic>=0.40.0,<1.0.0
voyageai>=0.3.0,<1.0.0
supabase>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
PyPDF2>=3.0.0,<4.0.0
tiktoken>=0.5.0,<1.0.0
tenacity>=8.0.0,<9.0.0
pydantic>=2.0.0,<3.0.0
pytest>=7.0.0,<8.0.0
ruff>=0.1.0

# Security API Layer (added for Backend-First architecture)
fastapi>=0.109.0,<1.0.0
uvicorn>=0.27.0,<1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
httpx>=0.26.0,<1.0.0
